pyproject.toml
[tool.poetry]
name = "article-view-tracker"
version = "0.1.0"
description = "Backend service for tracking article views"
authors = ["Vlad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
redis = "^5.2"
django-redis = "^5.4"
celery = "^5.4"
gunicorn = "^23.0"
django-environ = "^0.11"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"
factory-boy = "^3.3"
black = "^24.0"
isort = "^5.13"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"